Cleaning up tables when they are deleted, refactored a bit to make more similar to table save.
